% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen_Class.R
\name{data.gen.circles}
\alias{data.gen.circles}
\title{Circles}
\usage{
data.gen.circles(
  n,
  r_vec = c(1, 2),
  start = runif(1, -1, 1),
  s,
  do.plot = TRUE
)
}
\arguments{
\item{n}{The data length to be generated.}

\item{r_vec}{The radius of circles.}

\item{start}{The center of circles.}

\item{s}{The level of Gaussian noise, default 0.}

\item{do.plot}{Logical value. If TRUE (default value), a plot of the generated Circles is shown.}
}
\value{
A list of two variables, x and classes.
}
\description{
Circles
}
\examples{
Circles=data.gen.circles(n = 1000, r_vec=c(1,2), start=runif(1,-1,1), s=0.1, do.plot=TRUE)
}
