% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen_NL.R
\name{data.gen.nl1}
\alias{data.gen.nl1}
\title{Nonlinear system with independent/correlate covariates}
\usage{
data.gen.nl1(nobs, ndim = 15, r = 0.6, noise = 1)
}
\arguments{
\item{nobs}{The data length to be generated.}

\item{ndim}{The number of potential predictors (default is 9).}

\item{r}{Target Spearman correlation among covariates.}

\item{noise}{The noise level in the time series.}
}
\value{

}
\description{
Nonlinear system with independent/correlate covariates
}
\examples{

}
