% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen_NL.R
\name{data.gen.nl2}
\alias{data.gen.nl2}
\title{Nonlinear system with Exogenous covariates}
\usage{
data.gen.nl2(nobs, ndim = 7, noise = 1)
}
\arguments{
\item{nobs}{The data length to be generated.}

\item{ndim}{The number of potential predictors (default is 9).}

\item{noise}{The noise level in the time series.}
}
\value{

}
\description{
Nonlinear system with Exogenous covariates
}
\examples{

}
\references{
Sharma, A., & Mehrotra, R. (2014). An information theoretic alternative to model a natural system using observational information alone. Water Resources Research, 50(1), 650-660.
}
