% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{data.gen.norm}
\alias{data.gen.norm}
\title{Generate correlated normal variates}
\usage{
data.gen.norm(n, mu = rep(0, 2), sd = rep(1, 2), r = 0.6, sigma)
}
\arguments{
\item{n}{The data length to be generated.}

\item{mu}{A vector giving the means of the variables.}

\item{sd}{A vector giving the standard deviation of the variables.}

\item{r}{The target Pearson correlation, default is 0.6.}

\item{sigma}{A positive-definite symmetric matrix specifying the covariance matrix of the variables.}
}
\value{
A matrix of correlated normal variates
}
\description{
Generate correlated normal variates
}
