% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen_Class.R
\name{data.gen.spirals}
\alias{data.gen.spirals}
\title{Spirals}
\usage{
data.gen.spirals(n, cycles = 1, s = 0, do.plot = TRUE)
}
\arguments{
\item{n}{The data length to be generated.}

\item{cycles}{The number of cycles of spirals.}

\item{s}{The level of Gaussian noise, default 0.}

\item{do.plot}{Logical value. If TRUE (default value), a plot of the generated Spirals is shown.}
}
\value{
A list of two variables, x and classes.
}
\description{
Spirals
}
\examples{
Spirals=data.gen.spirals(n = 2000, cycles=2, s=0.01, do.plot=TRUE)
}
\references{
Friedrich Leisch & Evgenia Dimitriadou (2010). mlbench: Machine Learning Benchmark Problems. R package version 2.1-1.
}
