% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen_SW.R
\name{data.gen.SW}
\alias{data.gen.SW}
\title{Generate predictor and response data: sinusoidal model}
\usage{
data.gen.SW(nobs = 500, freq = 50, A = 2, phi = pi, mu = 0, sd = 1)
}
\arguments{
\item{nobs}{The data length to be generated.}

\item{freq}{The frequencies in the generated response. Default freq=50.}

\item{A}{The amplitude of the sinusoidal series}

\item{phi}{The phase of the sinusoidal series}

\item{mu}{The mean of Gaussian noise in the variable.}

\item{sd}{The standard deviation of Gaussian noise in the variable.}
}
\value{
A list of time and x.
}
\description{
Generate predictor and response data: sinusoidal model
}
\examples{
sample=500
sw <- data.gen.SW(nobs=sample, freq = 25, A = 2, phi = 0.6*pi, mu=0, sd = 0.1)
plot(sw$t,sw$x, type='o', ylab='Cosines', xlab="t")
}
\references{
Shumway, R. H., & Stoffer, D. S. (2011). Characteristics of Time Series. In D. S. Stoffer (Ed.), Time series analysis and its applications (pp. 8-14). New York : Springer.
}
