% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen_SW.R
\name{data.gen.SW1}
\alias{data.gen.SW1}
\title{Generate predictor and response data: sinusoidal model}
\usage{
data.gen.SW1(nobs = 512, fp = 25, fd, sd.x = 0.1, sd.y = 0.1)
}
\arguments{
\item{nobs}{The data length to be generated.}

\item{fp}{The frequencies in the generated response.}

\item{fd}{A vector of frequencies for potential predictors. fd = c(3,5,10,15,25,30,55,70,95) used in the WRR paper.}

\item{sd.x}{The noise level in the predictor.}

\item{sd.y}{The noise level in the response.}
}
\value{
A list of 3 elements: a vector of response (x), a matrix of potential predictors (dp) with each column containing one potential predictor, and a vector of true predictor numbers.
}
\description{
Generate predictor and response data: sinusoidal model
}
