% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen_Class.R
\name{data.gen.blobs}
\alias{data.gen.blobs}
\title{Gaussian Blobs}
\usage{
data.gen.blobs(
  nobs = 100,
  features = 2,
  centers = 3,
  sd = 1,
  bbox = c(-10, 10),
  do.plot = TRUE
)
}
\arguments{
\item{nobs}{The data length to be generated.}

\item{features}{Features of dataset}

\item{centers}{Either the number of centers, or a matrix of the chosen centers}

\item{sd}{The level of Gaussian noise, default 1.}

\item{bbox}{The bounding box of the dataset.}

\item{do.plot}{Logical value. If TRUE (default value), a plot of the generated Blobs is shown.}
}
\value{
A list of two variables, x and classes.
}
\description{
Gaussian Blobs
}
\examples{
Blobs=data.gen.blobs(nobs=1000, features=2, centers=3, sd=1, bbox=c(-10,10), do.plot=TRUE)
}
