% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen_SW.R
\name{data.gen.SW}
\alias{data.gen.SW}
\title{Generate predictor and response data: sinusoidal model}
\usage{
data.gen.SW(nobs = 512, fp = 50, fd = 50, sd = 1)
}
\arguments{
\item{nobs}{The data length to be generated.}

\item{fp}{The frequencies in the generated response. Default fp=50.}

\item{fd}{A vector of frequencies for potential predictors. Default fd=50.}

\item{sd}{The noise level in the predictor.}
}
\value{
A list of 3 elements: a vector of response (x), a matrix of potential predictors (dp) with each column containing one potential predictor, and a vector of true predictor numbers.
}
\description{
Generate predictor and response data: sinusoidal model
}
\references{
Shumway, R. H., & Stoffer, D. S. (2011). Characteristics of Time Series. In D. S. Stoffer (Ed.), Time series analysis and its applications (pp. 8-14). New York : Springer.
}
